of: Use the standards compliant default address cells value
The default address cells value is 1.

According to the devicetree specification and the OpenFirmware standard
(IEEE 1275-1994) this value should be 2.

The device tree compiler already set this value to 2 and the powerpc
PROM code also use 2 as default value.
Moderm implementation should have the #address-cells property set and
should not rely on this default value.

Be consistent with standards and update the default address cells value.

Signed-off-by: Herve Codina <herve.codina@bootlin.com>